# swagger.yml
openapi: 3.0.0
paths:
...
  /api/items:
    get:
      operationId: ItemsController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: search
          required: false
          in: query
          description: Search by item name/code
          schema:
            type: string
        - name: name
          required: false
          in: query
          description: Search by item name/code
          schema:
            type: string
        - name: categoryCode
          required: false
          in: query
          schema:
            example: ANALGESICS
            type: string
        - name: sortBy
          required: false
          in: query
          schema:
            default: name
            type: string
            enum:
              - name
              - code
              - createdAt
        - name: sortOrder
          required: false
          in: query
          schema:
            default: asc
            type: string
            enum:
              - asc
              - desc
      responses:
        '200':
          description: ''
      security: &ref_5
        - bearer: []
      summary: List items with pagination, filtering and sorting
      tags: &ref_6
        - items
    post:
      operationId: ItemsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponseDto'
      security: *ref_5
      summary: Create a catalog item
      tags: *ref_6
  /api/items/dependencies/categories:
    get:
      operationId: ItemsController_listCategories
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: search
          required: false
          in: query
          description: Search by name or code
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_5
      summary: Search item categories by name or code
      tags: *ref_6
  /api/items/dependencies/generic-products:
    get:
      operationId: ItemsController_listGenericProducts
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: search
          required: false
          in: query
          description: Search by name or code
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_5
      summary: Search generic products by name or code
      tags: *ref_6
  /api/items/dependencies/uoms:
    get:
      operationId: ItemsController_listUoms
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: search
          required: false
          in: query
          description: Search by name or code
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_5
      summary: Search UOMs by name or code
      tags: *ref_6
  /api/items/{itemId}:
    get:
      operationId: ItemsController_get
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponseDto'
      security: *ref_5
      summary: Get item details by id
      tags: *ref_6
    put:
      operationId: ItemsController_replace
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponseDto'
      security: *ref_5
      summary: Create a catalog item
      tags: *ref_6
    patch:
      operationId: ItemsController_patch
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchItemDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponseDto'
      security: *ref_5
      summary: Partially update a catalog item
      tags: *ref_6
...
  /api/price-lists/{priceListId}/items/{priceListItemId}:
    patch:
      operationId: PricingController_updateItem
      parameters:
        - name: priceListId
          required: true
          in: path
          schema:
            type: string
        - name: priceListItemId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePriceListItemDto'
      responses:
        '200':
          description: ''
      security: *ref_17
      tags: *ref_18
...
  /api/categories:
    get:
      operationId: CategoriesController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: search
          required: false
          in: query
          description: Search query
          schema:
            type: string
        - name: sortBy
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: sortOrder
          required: false
          in: query
          schema:
            default: desc
            type: string
            enum:
              - asc
              - desc
        - name: filter
          required: false
          in: query
          description: Generic filter string
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: &ref_19
        - bearer: []
      summary: List categories with pagination/filter/sort
      tags: &ref_20
        - categories
    post:
      operationId: CategoriesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: ''
      security: *ref_19
      tags: *ref_20
  /api/categories/export:
    get:
      operationId: CategoriesController_export
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: search
          required: false
          in: query
          description: Search query
          schema:
            type: string
        - name: sortBy
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: sortOrder
          required: false
          in: query
          schema:
            default: desc
            type: string
            enum:
              - asc
              - desc
        - name: filter
          required: false
          in: query
          description: Generic filter string
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
  /api/categories/{categoryId}:
    put:
      operationId: CategoriesController_replace
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
    patch:
      operationId: CategoriesController_patch
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
    get:
      operationId: CategoriesController_get
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
    delete:
      operationId: CategoriesController_remove
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
...
  /api/website/categories/{slug}:
    get:
      operationId: WebsiteController_getCategoryBySlug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get category by slug with products
      tags: *ref_45
...
info:
  title: RxSoft Backend API
  description: Phase 3 identity module and API contracts
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Object:
      type: object
      properties: {}
...
    ItemCategoryResponse:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
      required:
        - id
        - code
        - name
    PharmaceuticsResponse:
      type: object
      properties:
        code:
          type: string
        clinicalName:
          type: string
        drugClass:
          type: string
        pharmaceutics:
          type: string
      required:
        - code
        - clinicalName
        - drugClass
        - pharmaceutics
    GenericProductResponse:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        pharmaceutics:
          $ref: '#/components/schemas/PharmaceuticsResponse'
        isPrescriptionRequired:
          type: boolean
        isControlledSubstance:
          type: boolean
      required:
        - id
        - code
        - name
        - pharmaceutics
        - isPrescriptionRequired
        - isControlledSubstance
    ItemResponseDto:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        categoryId:
          type: string
        genericProductId:
          type: string
        category:
          $ref: '#/components/schemas/ItemCategoryResponse'
        genericProduct:
          $ref: '#/components/schemas/GenericProductResponse'
        barcode:
          type: object
          nullable: true
        baseUomId:
          type: string
        purchaseUomId:
          type: object
          nullable: true
        saleUomId:
          type: object
          nullable: true
        saleUom:
          type: object
          nullable: true
        baseUom:
          type: object
          nullable: true
        purchaseUom:
          type: object
          nullable: true
        trackLot:
          type: boolean
        trackExpiry:
          type: boolean
        shelfLifeDays:
          type: object
          nullable: true
        isActive:
          type: boolean
        imageUrl:
          type: object
          nullable: true
        smallImageUrl:
          type: object
          nullable: true
        mediumImageUrl:
          type: object
          nullable: true
        largeImageUrl:
          type: object
          nullable: true
      required:
        - id
        - code
        - name
        - categoryId
        - genericProductId
        - category
        - genericProduct
        - barcode
        - baseUomId
        - purchaseUomId
        - saleUomId
        - saleUom
        - baseUom
        - purchaseUom
        - trackLot
        - trackExpiry
        - shelfLifeDays
        - isActive
        - imageUrl
        - smallImageUrl
        - mediumImageUrl
        - largeImageUrl
    CreateItemDto:
      type: object
      properties:
        code:
          type: string
          example: PCM001
        name:
          type: string
          example: Paracetamol 500mg Tablet
        categoryId:
          type: string
          example: 2f9e3c8a-7fb2-4f43-98b7-5f2a0d4f9021
        genericProductId:
          type: string
          example: 5c0d6f5d-2952-4924-8ec6-c2c488a54cab
        baseUomId:
          type: string
          example: 8f6c4d74-607a-4d90-96f5-4aa6f4a32b31
        purchaseUomId:
          type: string
          example: 8f6c4d74-607a-4d90-96f5-4aa6f4a32b31
        saleUomId:
          type: string
          example: 8f6c4d74-607a-4d90-96f5-4aa6f4a32b31
        barcode:
          type: string
          example: '1234567890123'
        trackLot:
          type: boolean
          default: true
        trackExpiry:
          type: boolean
          default: true
        shelfLifeDays:
          type: number
          example: 730
        isActive:
          type: boolean
          default: true
        imageUrl:
          type: string
        smallImageUrl:
          type: string
        mediumImageUrl:
          type: string
        largeImageUrl:
          type: string
      required:
        - code
        - name
        - categoryId
        - genericProductId
        - baseUomId
    PatchItemDto:
      type: object
      properties:
        isActive:
          type: boolean
          example: true
        name:
          type: string
          example: Paracetamol 500mg
        barcode:
          type: string
          example: '4901234567890'
...
    UpdatePriceListItemDto:
      type: object
      properties:
        itemId:
          type: string
        locationId:
          type: string
        currencyCode:
          type: string
        unitPrice:
          type: number
        startsAt:
          format: date-time
          type: string
        endsAt:
          type: string
...
    CreateCategoryDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        parentId:
          type: string
      required:
        - code
        - name
    UpdateCategoryDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        parentId:
          type: string
...